+++ src/main.rs
@@ -12,7 +12,7 @@ mod tui;
 mod util;
 
 use crate::{
-    cli::Subcommand, collection::RequestCollection, tui::Tui, util::Directory,
+    cli::CliCommand, collection::RequestCollection, tui::Tui, util::Directory,
 };
 use clap::Parser;
 use std::{fs::File, path::PathBuf};
@@ -26,15 +26,21 @@ use tracing_subscriber::{filter::EnvFilter, prelude::*};
     long_about = "Configurable HTTP client with both TUI and CLI interfaces"
 )]
 struct Args {
+    #[command(flatten)]
+    global: GlobalArgs,
+    /// Subcommand to execute. If omitted, run the TUI
+    #[command(subcommand)]
+    subcommand: Option<CliCommand>,
+}
+
+/// Arguments that are available to all subcommands and the TUI
+#[derive(Debug, Parser)]
+struct GlobalArgs {
     /// Collection file, which defines your profiless and recipes. If omitted,
     /// check for the following files in the current directory (first match
     /// will be used): slumber.yml, slumber.yaml, .slumber.yml, .slumber.yaml
     #[clap(long, short)]
     collection: Option<PathBuf>,
-
-    /// Subcommand to execute. If omitted, run the TUI
-    #[command(subcommand)]
-    subcommand: Option<Subcommand>,
 }
 
 #[tokio::main]
@@ -47,13 +53,14 @@ async fn main() -> anyhow::Result<()> {
     match args.subcommand {
         // Run the TUI
         None => {
-            let collection_path = RequestCollection::try_path(args.collection)?;
+            let collection_path =
+                RequestCollection::try_path(args.global.collection)?;
             Tui::start(collection_path).await?;
             Ok(())
         }
 
         // Execute one request without a TUI
-        Some(subcommand) => subcommand.execute(args.collection).await,
+        Some(subcommand) => subcommand.execute(args.global).await,
     }
 }
